feat: implemented functionality Start Simulation

+++ src/main/java/lapTimeSimulator/utils/dto/outputDataDTO/VehicleDataOutDTO.java
@@ -14,9 +14,9 @@ public class VehicleDataOutDTO {
     public final double mass;
     public final double powerMax;
     public final double torqueMax;
-    public final double rpmPowerMax;
-    public final double rpmTorqueMax;
-    public final int numberOfGears;
+    public final Double rpmPowerMax;
+    public final Double rpmTorqueMax;
+    public final double numberOfGears;
     public final List<Double> gears;
     public final double finalDriveRatio;
     public final double longitudinalGrip;
